set up auth route group and auth redirections between home page and auth pages
