+++ frontend/src/app/(auth)/_layout.tsx
@@ -1,0 +1,16 @@
+import { Redirect, Slot } from "expo-router";
+import { useContext } from "react";
+
+import { UserContext } from "@/contexts/userContext";
+
+const AuthLayout = () => {
+  const { firebaseUser } = useContext(UserContext);
+
+  if (firebaseUser) {
+    return <Redirect href="/" />;
+  }
+
+  return <Slot />;
+};
+
+export default AuthLayout;
