+++ frontend/src/app/_layout.tsx
@@ -7,8 +7,6 @@ const RootLayout = () => {
     <UserContextProvider>
       <Stack screenOptions={{ headerShown: false }}>
         <Stack.Screen name="index" />
-        <Stack.Screen name="login" />
-        <Stack.Screen name="loading" />
       </Stack>
     </UserContextProvider>
   );
