+++ frontend/src/app/index.tsx
@@ -1,22 +1,27 @@
+import { Redirect } from "expo-router";
+import { useContext } from "react";
 import { Image, ImageSourcePropType, StatusBar, StyleSheet, Text, View } from "react-native";
 
-import Button from "../components/Button";
+import logo from "../assets/Poc_Mascots.png";
 
-import logo from "@/assets/Poc_Mascots.png";
+import Button from "@/components/Button";
+import { UserContext } from "@/contexts/userContext";
+import { logout } from "@/lib/auth";
 
-export default function App() {
+export default function Loading() {
+  const { firebaseUser } = useContext(UserContext);
+
+  if (!firebaseUser) {
+    return <Redirect href="/landing" />;
+  }
+
   return (
     <View style={styles.container}>
-      <View style={styles.topSection}>
-        <Image source={logo as ImageSourcePropType} style={styles.logo} />
-        <Text style={styles.title}>Psyches of Color</Text>
-        <Text style={styles.text}>My melanated psyche is my superpower</Text>
-      </View>
-      <View style={styles.bottomSection}>
-        <Button href="/login">Login</Button>
-        <Button href="/signup">Sign Up</Button>
-      </View>
+      <Image source={logo as ImageSourcePropType} style={styles.logo} />
+      <Text style={styles.title}>Psyches of Color</Text>
+      <Text style={styles.text}>slogan/mindful tip goes here</Text>
       <StatusBar />
+      <Button onPress={() => void logout()}>Logout</Button>
     </View>
   );
 }
@@ -25,17 +30,6 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: "#fff",
-    // alignItems: "center",
-    // justifyContent: "center",
-    justifyContent: "space-between",
-  },
-  topSection: {
-    marginTop: 200,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  bottomSection: {
-    marginBottom: 25,
     alignItems: "center",
     justifyContent: "center",
   },
@@ -51,6 +45,7 @@ const styles = StyleSheet.create({
     fontStyle: "normal",
     fontWeight: "400",
     lineHeight: 25.5,
+    marginBottom: 16,
   },
   title: {
     color: "#000000",
